requirements.txt
pandas ~=1.3.4
# dash version which allows date as input field
https://github.com/the-rene/dash/archive/refs/heads/dev.zip
dash-datetimepicker ~= 0.0.6
mysql-connector-python ~=8.0
SQLAlchemy ~=1.4.26
tabulate ~=0.8.9
dash-bootstrap-components ~= 1.0.2
gunicorn~=20.1.0